refactor(client): apply PascalCase naming for component files and update imports
- Renamed multiple component files from kebab-case to PascalCase for consistency
- Updated relevant import paths across App.tsx, Home.tsx, Profile.tsx, Register.tsx, and component files
- Minor modifications in components: Navbar, LoginForm, RegisterForm, ProfileComponent, and routing logic

This improves code readability and aligns with React naming conventions.
